+++ munetic_express/src/controllers/user.controller.ts
@@ -1,7 +1,5 @@
 import { RequestHandler } from 'express';
 import * as Status from 'http-status';
-import multer from 'multer';
-import path from 'path';
 import ErrorResponse from '../modules/errorResponse';
 import { ResJSON } from '../modules/types';
 import * as UserService from '../service/user.service';
@@ -55,38 +53,6 @@ export const editUserProfile: RequestHandler = async (req, res, next) => {
   } catch (err) {
     next(err);
   }
-};
-
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, '../munetic_app/public/img');
-  },
-  filename: function (req, file, cb) {
-    const ext = path.extname(file.originalname);
-    cb(null, path.basename(file.originalname, ext) + '-' + Date.now() + ext);
-  },
-});
-
-const storageConfig = multer({
-  storage: storage,
-  limits: { fileSize: 1000000 },
-  fileFilter: function (req, file, callback) {
-    const ext = path.extname(file.originalname);
-    if (
-      ext !== '.png' &&
-      ext !== '.jpg' &&
-      ext !== '.jpeg' &&
-      ext !== '.gif' &&
-      ext !== '.svg'
-    ) {
-      return callback(new Error('이미지 형식이 잘못됐습니다.'));
-    }
-    callback(null, true);
-  },
-});
-
-export const imgUpload: RequestHandler = (req, res, next) => {
-  return storageConfig.single('img')(req, res, next);
 };
 
 export const createProfileImg: RequestHandler = async (req, res, next) => {
